Fix wrong commands path on Makefile
