+++ cmd/gitbase/command/server.go
@@ -1,4 +1,4 @@
-package cmd
+package command
 
 import (
 	"net"
